requirements.txt
django==5.2.3
sqlparse==0.5.3
django-bootstrap4==25.1
djangorestframework==3.16.0
factory_boy==3.3.3
psycopg2==2.9.10
pillow==11.3.0
celery==5.5.3
redis==6.4.0
channels==4.3.1
channels-redis==4.3.0
djangorestframework-simplejwt==5.5.1
daphne==4.2.1
django-htmx==1.23.2
hypothesis==6.138.3
pyyaml==6.0.2
uritemplate==4.2.0
inflection==0.5.1
drf-spectacular==0.28.0
python-dotenv==1.1.1